users and roles

+++ SaveIt/Controllers/CommentsController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using SaveIt.Data;
 using SaveIt.Models;
 
@@ -8,11 +10,18 @@ namespace SaveIt.Controllers
     {
         private readonly ApplicationDbContext db;
 
-        public CommentsController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        private readonly RoleManager<IdentityRole> _roleManager;
+
+        public CommentsController(ApplicationDbContext context, UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
         {
             db = context;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
 
+        /*
         [HttpPost]
         public IActionResult New(Comment comment)
         {
@@ -27,30 +36,51 @@ namespace SaveIt.Controllers
             {
                 return Redirect("/Pins/Show/" + comment.PinId);
             }
-        }
+        }*/
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult Delete(int id)
         {
             Comment comment = db.Comments.Find(id);
+            if (comment.UserId != _userManager.GetUserId(User) && !User.IsInRole("Admin"))
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti acest comentariu!";
+                TempData["messageType"] = "alert-danger";
+                return Redirect("/Pins/Show/" + comment.PinId);
+            }
             db.Comments.Remove(comment);
             db.SaveChanges();
             return Redirect("/Pins/Show/" + comment.PinId);
         }
 
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Edit(int id)
         {
             Comment comment = db.Comments.Find(id);
+            if (comment.UserId != _userManager.GetUserId(User) && !User.IsInRole("Admin"))
+            {
+                TempData["message"] = "Nu aveti dreptul sa editati acest comentariu!";
+                TempData["messageType"] = "alert-danger";
+                return Redirect("/Pins/Show/" + comment.PinId);
+            }
             ViewBag.Comment = comment;
             return View();
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult Edit(int id, Comment requestComment)
         {
             Comment comment = db.Comments.Find(id);
             if (ModelState.IsValid)
             {
+                if (comment.UserId != _userManager.GetUserId(User) && !User.IsInRole("Admin"))
+                {
+                    TempData["message"] = "Nu aveti dreptul sa editati acest comentariu!";
+                    TempData["messageType"] = "alert-danger";
+                    return Redirect("/Pins/Show/" + comment.PinId);
+                }
                 comment.Content = requestComment.Content;
                 db.SaveChanges();
                 return Redirect("/Pins/Show/" + comment.PinId);
