+++ SaveIt/Controllers/PinsController.cs
@@ -1,4 +1,6 @@
 ﻿using Humanizer;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -11,15 +13,22 @@ namespace SaveIt.Controllers
     {
         private readonly ApplicationDbContext db;
 
-        public PinsController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        private readonly RoleManager<IdentityRole> _roleManager;
+
+        public PinsController(ApplicationDbContext context, UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
         {
             db = context;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
 
+        //[Authorize(Roles = "User,Admin")]
         public IActionResult Index()
         {
             //var pins = db.Pins.Include("PinTags.Tag");
-            var pins = db.Pins.Include(p => p.PinTags).ThenInclude(pt => pt.Tag);
+            var pins = db.Pins.Include("User").Include(p => p.PinTags).ThenInclude(pt => pt.Tag);
             ViewBag.Pins = pins;
             
             if (TempData.ContainsKey("message"))
@@ -30,18 +39,27 @@ namespace SaveIt.Controllers
             return View();
         }
 
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Show(int id)
         {
             //Pin pin = db.Pins.Include("PinTags.Tag").Include("Comments").Where(p => p.Id == id).First();
-            Pin pin = db.Pins.Include(p => p.PinTags).ThenInclude(pt => pt.Tag).Include(p => p.Comments).FirstOrDefault(p => p.Id == id);
+            Pin pin = db.Pins.Include("User").Include(p => p.PinTags).ThenInclude(pt => pt.Tag).Include(p => p.Comments).Include("Comments.User").FirstOrDefault(p => p.Id == id);
+            SetAccessRights();
             return View(pin);
         }
 
+        private void SetAccessRights()
+        {
+            ViewBag.EsteAdmin = User.IsInRole("Admin");
+            ViewBag.UserCurent = _userManager.GetUserId(User);
+        }
+
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult Show([FromForm] Comment comment)
         {
             comment.Date = System.DateTime.Now;
-            //comment.UserId = _userManager.GetUserId();
+            comment.UserId = _userManager.GetUserId(User);
             if (ModelState.IsValid)
             {
                 db.Comments.Add(comment);
@@ -51,7 +69,8 @@ namespace SaveIt.Controllers
             else
             {
                 //Pin pin = db.Pins.Include("PinTags.Tag").Include("Comments").Where(p => p.Id == comment.PinId).First();
-                Pin pin = db.Pins.Include(p => p.PinTags).ThenInclude(pt => pt.Tag).Include(p => p.Comments).FirstOrDefault(p => p.Id == comment.PinId);
+                Pin pin = db.Pins.Include("User").Include(p => p.PinTags).ThenInclude(pt => pt.Tag).Include(p => p.Comments).Include("Comments.User").FirstOrDefault(p => p.Id == comment.PinId);
+                SetAccessRights();
                 return View(pin);
             }
         }
@@ -73,6 +92,7 @@ namespace SaveIt.Controllers
             return selectList;
         }
 
+        [Authorize(Roles = "User,Admin")]
         public IActionResult New()
         {
             Pin pin = new Pin();
@@ -80,12 +100,12 @@ namespace SaveIt.Controllers
             return View(pin);
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult New(Pin pin)
         {
             pin.Date = DateTime.Now;
-            pin.Tags = GetAllTags();
-            //pin.UserId = _userManager.GetUserId();
+            pin.UserId = _userManager.GetUserId(User);
             if (ModelState.IsValid)
             {
                 db.Pins.Add(pin);
@@ -106,37 +126,50 @@ namespace SaveIt.Controllers
             }
             else
             {
+                pin.Tags = GetAllTags();
                 return View(pin);
             }
         }
-        
+
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Edit(int id)
         {
             //Pin pin = db.Pins.Include("PinTags.Tag").Where(art => art.Id == id).First();
-            Pin pin = db.Pins.Include(p => p.PinTags).ThenInclude(pt => pt.Tag).FirstOrDefault(p => p.Id == id);
+            Pin pin = db.Pins.Include("User").Include(p => p.PinTags).ThenInclude(pt => pt.Tag).FirstOrDefault(p => p.Id == id);
             pin.Tags = GetAllTags();
+
+            if (pin.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(pin);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti drepturi pentru aceasta actiune!";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index");
+            }
 
             return View(pin);
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult Edit(int id, Pin requestPin)
         {
             if (ModelState.IsValid)
             {
                 Pin pin = db.Pins.Include(p => p.PinTags).Where(p => p.Id == id).First();
+                
+                if (pin.UserId != _userManager.GetUserId(User) && !User.IsInRole("Admin"))
+                {
+                    TempData["message"] = "Nu aveti drepturi pentru aceasta actiune!";
+                    TempData["messageType"] = "alert-danger";
+                    return RedirectToAction("Index");
+                }
+
                 pin.Title = requestPin.Title;
                 pin.Content = requestPin.Content;
                 pin.Date = DateTime.Now;
-                // trebuie pus si tag-ul
-                //pin.Tags = GetAllTags();
-                // vreau ca lista de TagIds sa fie copiata de la requestPin
-                //pin.TagIds = requestPin.TagIds;
-                //pin.TagId = requestPin.TagId;
-                //pin.Tags = GetAllTags();
-                //pin.PinTags = requestPin.PinTags;
-                //db.Pins.Update(pin);
-                //db.SaveChanges();
                 pin.PinTags.Clear();
                 foreach (var tagId in requestPin.TagIds)
                 {
@@ -159,10 +192,17 @@ namespace SaveIt.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult Delete(int id)
         {
-            var pin = db.Pins.Find(id);
+            var pin = db.Pins.Include("Comments").Where(p => p.Id == id).First();
+            if (pin.UserId != _userManager.GetUserId(User) && !User.IsInRole("Admin"))
+            {
+                TempData["message"] = "Nu aveti drepturi pentru aceasta actiune!";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index");
+            }
             db.Pins.Remove(pin);
             db.SaveChanges();
             TempData["message"] = "Pin-ul a fost sters!";
